feat(resultsHomeTableFiltering): Refactored filtering methods into separate file
