+++ src/Analysis/GWASResultsV2/Views/Home/HomeTable/filterTableData.js
@@ -1,0 +1,64 @@
+import moment from 'moment';
+
+const filterBySearchTerm = (initData, key, searchTerm) =>
+  initData.filter((obj) =>
+    obj[key]
+      .toString()
+      .toLowerCase()
+      .includes(searchTerm.toLowerCase())
+  );
+
+const filterByJobStatuses = (initData, homeTableState) =>
+  initData.filter((item) =>
+    homeTableState.jobStatusSelections.includes(item.phase)
+  );
+
+const filterByDateRange = (initData, key, dateSelection) =>
+  initData.filter((obj) => {
+    const utcDate = moment.utc(obj[key]);
+    return (
+      utcDate.isSameOrAfter(dateSelection[0]) &&
+      utcDate.isSameOrBefore(dateSelection[1])
+    );
+  });
+
+const filterTableData = (data, homeTableState) => {
+  let filteredDataResult = data;
+  if (homeTableState.nameSearchTerm.length > 0) {
+    filteredDataResult = filterBySearchTerm(
+      filteredDataResult,
+      'name',
+      homeTableState.nameSearchTerm
+    );
+  }
+  if (homeTableState.wfNameSearchTerm.length > 0) {
+    filteredDataResult = filterBySearchTerm(
+      filteredDataResult,
+      'wf_name',
+      homeTableState.wfNameSearchTerm
+    );
+  }
+  if (homeTableState.submittedAtSelections.length > 0) {
+    filteredDataResult = filterByDateRange(
+      filteredDataResult,
+      'submittedAt',
+      homeTableState.submittedAtSelections
+    );
+  }
+  if (homeTableState.jobStatusSelections.length > 0) {
+    filteredDataResult = filterByJobStatuses(
+      filteredDataResult,
+      homeTableState
+    );
+  }
+  if (homeTableState.startedAtSelections.length > 0) {
+    filteredDataResult = filterByDateRange(
+      filteredDataResult,
+      'startedAt',
+      homeTableState.startedAtSelections
+    );
+  }
+  return filteredDataResult;
+};
+
+export default filterTableData;
